tests fix + cleanup

+++ src/main/groovy/com/apphance/ameba/android/plugins/buildplugin/AndroidPlugin.groovy
@@ -392,16 +392,16 @@ class AndroidPlugin implements Plugin<Project> {
         task.dependsOn(project.readAndroidProjectConfiguration, project.verifySetup, project.copySources)
         project.tasks["buildAll${debugRelease}"].dependsOn(task)
     }
-	
+
 	void prepareCopySourcesTask(Project project) {
 		def task = project.task('copySources')
 		task.description = "Copies all sources to tmp directory for build"
 		task.group = AmebaCommonBuildTaskGroups.AMEBA_BUILD
 		task << {
-			new AntBuilder().delete(dir: "srcTmp")
-			new AntBuilder().copy(toDir : "srcTmp", verbose:true) {
-				fileset(dir : "./") {
-					exclude(name: 'srcTmp')
+			new AntBuilder().delete(dir: "${project.rootDir}/srcTmp")
+			new AntBuilder().copy(toDir : "${project.rootDir}/srcTmp", verbose:true) {
+				fileset(dir : "${project.rootDir}/") {
+					exclude(name: "${project.rootDir}/srcTmp/**/*")
 					conf.sourceExcludes.each {
 						if (!it.equals('**/local.properties')) {
 							exclude(name: it)
@@ -410,9 +410,11 @@ class AndroidPlugin implements Plugin<Project> {
 				}
 			}
 			// edit ant.properties for proper signing keystore path
-			File antProperties = new File('srcTmp/ant.properties')
-			File newAntProperties = new File('srcTmp/ant.props')
-			newAntProperties.delete()
+			File antProperties = new File("${project.rootDir}/srcTmp/ant.properties")
+			File newAntProperties = new File("${project.rootDir}/srcTmp/ant.props")
+			if (newAntProperties.exists()) {
+				newAntProperties.delete()
+			}
 			newAntProperties.withWriter { out ->
 				antProperties.eachLine {
 					if (it.contains("key.store=")) {
